requirements.txt
aiofiles==0.5.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
DAWG-Python==0.7.2
docopt==0.6.2
fastapi==0.55.1
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.3.0
packaging==20.4
peewee==3.13.3
pluggy==0.13.1
py==1.8.1
pydantic==1.5.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
pyparsing==2.4.7
pytest==5.4.2
requests==2.23.0
six==1.15.0
starlette==0.13.2
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
zipp==3.1.0
psycopg2-binary==2.8.3
psycopg2==2.8.5